Move metrics functions to metrics.go
